Create optihorni.py
this is just me working on logics so files will come and go merge with others and such have fun. this temporary file for completion which is main focus for merging into horns. you  will be able to choose a flair rate or let the calculations optimize the horn for your drivers Fh and Fl range so it will auto calculate the best flair rate.